feat(map): optimize map service and render map js script

+++ src/main/java/pl/czyzlowie/modules/map/dto/MapMarkerDto.java
@@ -16,8 +16,8 @@ public class MapMarkerDto {
 
     private String id;
     private String name;
-    private BigDecimal lat;
-    private BigDecimal lng;
+    private Double lat;
+    private Double lng;
     private String type;
     private String slug;
     private String description;
